return CustomAutoSchema to apps/iam, move ORGANIZATION_OPEN_API_PARAMETERS from apps/engine to apps/iam
