+++ cvat/apps/iam/filters.py
@@ -1,15 +1,43 @@
-# Copyright (C) 2021-2022 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
 
 from rest_framework.filters import BaseFilterBackend
 
+from drf_spectacular.utils import OpenApiParameter
+
+ORGANIZATION_OPEN_API_PARAMETERS = [
+    OpenApiParameter(
+        name='org',
+        type=str,
+        required=False,
+        location=OpenApiParameter.QUERY,
+        description="Organization unique slug",
+    ),
+    OpenApiParameter(
+        name='org_id',
+        type=int,
+        required=False,
+        location=OpenApiParameter.QUERY,
+        description="Organization identifier",
+    ),
+    OpenApiParameter(
+        name='X-Organization',
+        type=str,
+        required=False,
+        location=OpenApiParameter.HEADER,
+        description="Organization unique slug",
+    ),
+]
+
 class OrganizationFilterBackend(BaseFilterBackend):
-    organization_slug = 'org'
-    organization_slug_description = 'Organization unique slug'
-    organization_id = 'org_id'
-    organization_id_description = 'Organization identifier'
-    organization_slug_header = 'X-Organization'
+
+    def _parameter_is_provided(self, request):
+        for parameter in ORGANIZATION_OPEN_API_PARAMETERS:
+            if parameter.location == 'header' and parameter.name in request.header:
+                return True
+            elif parameter.location == 'query' and parameter.name in request.query_params:
+                return True
+        return False
 
     def filter_queryset(self, request, queryset, view):
         # Filter works only for "list" requests and allows to return
@@ -24,11 +52,7 @@ class OrganizationFilterBackend(BaseFilterBackend):
         if org:
             visibility = {'organization': org.id}
 
-        elif not org and (
-            self.organization_slug in request.query_params
-            or self.organization_id in request.query_params
-            or self.organization_slug_header in request.headers
-        ):
+        elif not org and self._parameter_is_provided(request):
             visibility = {'organization': None}
 
         if visibility:
@@ -41,23 +65,20 @@ class OrganizationFilterBackend(BaseFilterBackend):
         if not view.iam_organization_field or view.detail:
             return []
 
-        return [
-            {
-                'name': self.organization_slug,
-                'in': 'query',
-                'description': self.organization_slug_description,
-                'schema': {'type': 'string'},
-            },
-            {
-                'name': self.organization_id,
-                'in': 'query',
-                'description': self.organization_id_description,
-                'schema': {'type': 'integer'},
-            },
-            {
-                'name': self.organization_slug_header,
-                'in': 'header',
-                'description': self.organization_slug_description,
-                'schema': {'type': 'string'},
-            },
-        ]
+        parameters = []
+        for parameter in ORGANIZATION_OPEN_API_PARAMETERS:
+            parameter_type = None
+
+            if parameter.type == int:
+                parameter_type = 'integer'
+            elif parameter.type == str:
+                parameter_type = 'string'
+
+            parameters.append({
+                'name': parameter.name,
+                'in': parameter.location,
+                'description': parameter.description,
+                'schema': {'type': parameter_type}
+            })
+
+        return parameters
