+++ cvat/settings/base.py
@@ -188,7 +188,7 @@ REST_FRAMEWORK = {
         'anon': '100/minute',
     },
     'DEFAULT_METADATA_CLASS': 'rest_framework.metadata.SimpleMetadata',
-    'DEFAULT_SCHEMA_CLASS': 'cvat.apps.engine.schema.CustomAutoSchema',
+    'DEFAULT_SCHEMA_CLASS': 'cvat.apps.iam.schema.CustomAutoSchema',
     'EXCEPTION_HANDLER': 'cvat.apps.events.handlers.handle_viewset_exception',
 }
